rm test file
